msm: kgsl: Update the sequence to resume stalled SMMU
On MMU500, we need to clear FSR first and the resume
SMMU to make sure FSR stalled status bit is cleared.

Also, as we are clearing FSR first SCTLR.HUPCF doesn't
take care of terminating any pending transaction. To
handle this run above sequence in loop till all pending
transaction on SMMU side are terminated and FSR.SS bit
is cleared.

This logic works for QSMMU as well.

Change-Id: Ie9b4cbe896fa0d5be487c07425150582569edbd9
Signed-off-by: Deepak Kumar <dkumar@codeaurora.org>
Signed-off-by: Dušan Uverić <dusan.uveric9@gmail.com>